requirements.txt
playwright==1.40.0
allure-pytest==2.13.2
python-dotenv==1.0.0
PyYAML==6.0
